requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
apscheduler
httpx
faiss-cpu
streamlit
email-validator
defusedxml
